Reduces boilerplate when creating server. Aditionally: * Unifies test utilities to DRY among @tagless and @service tests * Removes ETA Expansion warnings * Removes bootstrapFuture syntax * Removes all those imports that are not used anymore * Others
