+++ rpc/src/test/scala/client/handlers/TaskMHandlerTests.scala
@@ -35,7 +35,6 @@ class TaskMHandlerTests extends RpcClientTestSuite {
       "freestyle.async.AsyncContext[F] is provided" in {
 
       implicit val S: monix.execution.Scheduler = monix.execution.Scheduler.Implicits.global
-      import freestyle.async.implicits.futureAsyncContext
 
       val handler: Task ~> Future = task2Future
       val fooTask: Task[String]   = Task.now(foo)
@@ -46,7 +45,6 @@ class TaskMHandlerTests extends RpcClientTestSuite {
     "recover from a failed monix.eval.Task wrapping them into scala.concurrent.Future" in {
 
       implicit val S: monix.execution.Scheduler = monix.execution.Scheduler.Implicits.global
-      import freestyle.async.implicits.futureAsyncContext
 
       val handler: Task ~> Future  = task2Future
       val taskFailed: Task[String] = Task.raiseError(new RuntimeException(failureMessage))
