+++ rpc/src/test/scala/server/RpcServerTestSuite.scala
@@ -42,17 +42,23 @@ trait RpcServerTestSuite extends RpcBaseTestSuite {
     val immutableServiceList: List[ServerServiceDefinition] = List(sd1)
     val mutableServiceList: List[ServerServiceDefinition]   = List(sd2)
 
-    (serverMock.start _).when().returns(serverCopyMock)
-    (serverMock.getPort _).when().returns(port)
-    (serverMock.getServices _).when().returns(serviceList.asJava)
-    (serverMock.getImmutableServices _).when().returns(immutableServiceList.asJava)
-    (serverMock.getMutableServices _).when().returns(mutableServiceList.asJava)
-    (serverMock.shutdown _).when().returns(serverCopyMock)
-    (serverMock.shutdownNow _).when().returns(serverCopyMock)
-    (serverMock.isShutdown _).when().returns(b)
-    (serverMock.isTerminated _).when().returns(b)
+    (serverMock.start _: () => Server).when().returns(serverCopyMock)
+    (serverMock.getPort _: () => Int).when().returns(port)
+    (serverMock.getServices _: () => java.util.List[ServerServiceDefinition])
+      .when()
+      .returns(serviceList.asJava)
+    (serverMock.getImmutableServices _: () => java.util.List[ServerServiceDefinition])
+      .when()
+      .returns(immutableServiceList.asJava)
+    (serverMock.getMutableServices _: () => java.util.List[ServerServiceDefinition])
+      .when()
+      .returns(mutableServiceList.asJava)
+    (serverMock.shutdown _: () => Server).when().returns(serverCopyMock)
+    (serverMock.shutdownNow _: () => Server).when().returns(serverCopyMock)
+    (serverMock.isShutdown _: () => Boolean).when().returns(b)
+    (serverMock.isTerminated _: () => Boolean).when().returns(b)
     (serverMock.awaitTermination(_: Long, _: TimeUnit)).when(timeout, timeoutUnit).returns(b)
-    (serverMock.awaitTermination _).when().returns(unit)
+    (serverMock.awaitTermination _: () => Unit).when().returns(unit)
   }
 
   object implicits extends Helpers with DummyData {
